Implementando rotacao

+++ parser.c
@@ -7,7 +7,7 @@
 #define MAX_NAME_SIZE 100
 #define MAX 100
 
-void define_io(int argc, char **argv, char ** input, char ** output, float * angulo, float * limiar, int * mediana)
+void define_io(int argc, char **argv, char ** input, char ** output, double * angulo, float * limiar, int * mediana)
 {
 	int option;
 	char * angulo_char = NULL;
