+++ parser.h
@@ -1,2 +1,2 @@
-void define_io(int argc, char **argv, char ** input, char ** output, float * angulo, float * limiar, int * mediana);
+void define_io(int argc, char **argv, char ** input, char ** output, double * angulo, float * limiar, int * mediana);
